apiAyuda/User/views.py
from django.shortcuts import render

# Create your views here.

from rest_framework.views import APIView
from rest_framework import status
from rest_framework.response import Response
from rest_framework.permissions import AllowAny

from .serializers import UserSerializer
from .models import User


class UsersRegisterApi(APIView):

    permission_clases=(AllowAny,)

    def get(self, request):
        personas=User.objects.all()
        serializer=UserSerializer(personas, many=True)
        return Response(serializer.data,status=status.HTTP_200_OK)

    def post(self, request):
        serializer=UserSerializer(data=request.data)
        if serializer.is_valid():
            serializer.save()
            return Response(serializer.data, status=status.HTTP_201_CREATED)
        else:
            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)